requirements.txt
paddlepaddle
paddleocr
opencv-python
numpy
pymupdf
python-docx
pillow
tqdm
